bindings/cpp: Add FileStream
It's a wrapper for `ouisync::File` which implements asio's `AsyncReadStream` (`AsyncWriteStream` is
TODO). The bulk of the code taken from ouinet's `OuisyncFile` and modified to use generic completion
token instead of `yield_context`.

+++ bindings/cpp/tests/test_utils.hpp
@@ -77,18 +77,19 @@ public:
     }
 
     ~TempDir() {
-        boost::filesystem::remove_all(_path);
+        if (!_path.empty()) {
+            boost::filesystem::remove_all(_path);
+        }
     }
 
     TempDir(const TempDir&) = delete;
-    TempDir(TempDir&&) = delete;
+    TempDir(TempDir&&) = default;
     TempDir& operator=(const TempDir&) = delete;
-    TempDir& operator=(TempDir&&) = delete;
+    TempDir& operator=(TempDir&&) = default;
 
     const boost::filesystem::path& path() const {
         return _path;
     }
-
 };
 
 void check_exception(std::exception_ptr e);
